Moved It's a Trap to own upgrade.
