+++ Assets/Scripts/Model/Content/SecondEdition/Pilots/Rebel/ASF01BWing/GinaMoonsongBattleOverEndor.cs
@@ -133,44 +133,6 @@ namespace Abilities.SecondEdition
         }
     }
 
-    public class ItsATrapAbility : GenericAbility
-    {
-        public override void ActivateAbility()
-        {
-            HostShip.OnAttackStartAsDefender += RegisterAbility;
-        }
-
-        public override void DeactivateAbility()
-        {
-            HostShip.OnAttackStartAsDefender -= RegisterAbility;
-        }
-
-        private void RegisterAbility()
-        {
-            AddDiceModification(
-                "It's a Trap! ability",
-                IsDiceModificationAvailable,
-                GetDiceModificationAiPriority,
-                DiceModificationType.Reroll,
-                1,
-                new List<DieSide> { DieSide.Blank }
-            );
-        }
-
-        private bool IsDiceModificationAvailable()
-        {
-            List<GenericShip> friendlyShipsInRangeOne = Board.GetShipsAtRange(HostShip, new Vector2(0, 1), Team.Type.Friendly).Where(ship => ship != HostShip).ToList();
-            List<GenericShip> enemyShipsInRangeOne = Board.GetShipsAtRange(HostShip, new Vector2(0, 1), Team.Type.Enemy);
-
-            return friendlyShipsInRangeOne.Count > enemyShipsInRangeOne.Count && Combat.DiceRollDefence.Blanks > 0;
-        }
-
-        private int GetDiceModificationAiPriority()
-        {
-            return 1000;
-        }
-    }
-
     public class GyroCockpit : GenericAbility
     {
         // After you gain a stress token, you may spend 2 charges to gain an evade token.
@@ -274,22 +236,4 @@ namespace Abilities.SecondEdition
     }
 }
 
-namespace UpgradesList.SecondEdition
-{
-    public class ItsATrap : GenericUpgrade
-    {
-        public ItsATrap() : base()
-        {
-            UpgradeInfo = new UpgradeCardInfo(
-                "It's a Trap!",
-                UpgradeType.Talent,
-                cost: 0,
-                abilityType: typeof(ItsATrapAbility)
-            );
-
-            IsHidden = true;
-
-            ImageUrl = HostShip != null ? HostShip.ImageUrl : "https://infinitearenas.com/xw2/images/quickbuilds/ginamoonsong-battleoverendor.png";
-        }
-    }
 }
