+++ Assets/Scripts/Model/Content/SecondEdition/Upgrades/Talent/ItsATrap.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 6cb7183f813e8104b9be2ecdf50f068b
